Move setSessionVars to Unzer Service
